rm unnecessary logical_scope_guard tests
now that logical_scope_guard composes of logical_cleanup and aggregates
logical_init, many of its tests are not necessary. some have already
moved to test/logical_init and test/logical_cleanup, but for what i see
the rest are not necessary.

for example, consider testing the constructor

 struct logical_scope_guard{
   logical_scope_guard(logical_init, logical_cleanup);
   // ...
 };

the logical_scope_guard tests this constructor by listing a number of
tests: a test for the behaviour of the logical_init argument, a test for
the behaviour of the logical_cleanup argument, and a test for the
behaviour of both together. this is many tests and a lot of work. we've
already tested each of them separately, so why now test them again?

it would make sense to test how they work together, but i don't think
that's what these tests do. they really test individual work in the
presence of other objects, when the objects are (for what i see)
completely decoupled, unaware of each other.