cmake: differentiate among header files
there are public header files and private header files. within each of
these, there are forward declaration files, class definition files, and
template function definition files. among these, there is the tls::
stuff, and there is the tls::dim:: stuff.

the dim:: stuff is in directory dim/, and i want to this directory
strucure to remain when i install the software as well, so i have to add
install commands for dim files separate from the tls files.