fix dist regeneration
 $ touch test/main.cc && ninja dist

did not re-generate the dist files because the dist_* commands only
depended on check. check does not output anything, and so once cmake
finished with check there was no reason to also run the dist commands.

to fix, the dist commands must have an input that changes in order to
trigger re-generation of the dist files. so, make them depend on
check_sdl_toolset. then, recompiling the tests will cause a change in
one of the the dist files' inputs, and they will re-generate.