rerun tests only if dirty or they failed last time
if the tests pass, the command generates a tests_passed file to let
further builds know the tests have passed last time.

the command however now depends on the unit tests executable, so if the
unit tests executable is re-generated, the command will re-run the
tests, regardless of whether there is a tests_passed file or not.