translate_keydown_event()
for now just handle keys 0 and 1 being pressed down, to illustrate the
direction.

there's coming a translate_event() function that calls
translate_keydown_event() when the event is that of a key being pressed
down. translate_event() will similarly call a translation function for
mouse motion events, and so on. the event_loop() function will call the
translate_event() function, thereby kicking off the translation process
of an SDL event into what i prefer to work with: a boost signal.