push quit event: rm SDL_zero() call
i dont't think the call is necessary: in a quit event, no event data
matters other than the type. if i forget anything and there is anything
else in the event that matters then i should set it. but there is no
need to do the work of zeroing out the whole thing.

also, i learned SDL_zero() is a macro, and i don't feel like mocking
that in the upcoming sdl mock library...