extract key{down,up}_event_dispatch
event_dispatch composes of

 * keydown_event_dispatch
 * keyup_event_dispatch

to break it down to smaller chunks of logic.

keydown_event_dispatch in turn composes of

 * repeat_keydown_event_dispatch
 * nonrepeat_keydown_event_dispatch

finally, keydown_event_dispatch introduces keydown event handling that
is either a repeat or a nonrepeat. this makes registering to keydown
events more expressive when we want both repeat and nonrepeats, and it
also looks more symmetric with registering for keyup events.