event_dispatch signals take sdl_event
rather than translate sdl events to something else, let's simplify
things here. event_dispatch is now a thinner layer on top of sdl events:
we peek at the event type specified by the type field of the SDL_Event
union, and translate that to an event_dispatch signal. i don't
distinguish between events based on anything else. for example, there is
no event for pressing the left-arrow key and an event for pressing the
right-arrow key, an event for pressing the a key and holding the key
down and an event for pressing the a key and immediately releasing, etc.
rather, there's one keydown event; to get further information there's
the sdl event argument for all the details.

this way, the client will do the processing. i think this saves a lot of
headache on glasses' side, and also allows for flexibility on the client
side.