at least one update per draw
as it were, event_loop() could (and usually did, at least on the first
iteration) call draw() before calling update().

this might be okay, or might not, depending on the game code. either
way, i'm not sure if it's of any value to call draw before calling
update, or calling draw twice in a row.

let's guarantee an update will always precede a draw.